requirements.txt
pandas==1.3.5
torchvision==0.12.0
